feat: move verifyBlockStateTransition() and importBlock() to BeaconEngine
